Group 18/CarePlus+/requirements.txt
tkcalendar==1.6.1
mysql-connector-python==8.0.33
Pillow==10.2.0 